+++ tests/unit/test_crawlers.py
@@ -1,9 +1,10 @@
-from mock import mock_open, patch
+import time
+from mock import mock_open, patch, call, Mock
 
 from nose.tools import eq_
 
 from whynot.default_settings import DATABANKS
-from whynot.crawlers import DirCrawler, FileCrawler
+from whynot.crawlers import DirCrawler, FileCrawler, crawl
 
 
 @patch('os.path.exists', return_value=True)
@@ -40,3 +41,29 @@ def test_file_crawler(mock_exists, mock_getmtime):
         entries = FileCrawler.crawl(pdbfinder_databank['source'],
                                     pdbfinder_databank['regex'])
         eq_(len(entries), 2)
+
+
+@patch('whynot.crawlers.storage')
+def test_crawl(mock_storage):
+    entries = [
+        {
+            'pdb_id': '1crn', 'file_path': '/srv/data/pdb/flat/pdb1crn.ent',
+            'mtime': time.time(), 'comment': None,
+        },
+        {
+            'pdb_id': '2ltw', 'file_path': '/srv/data/pdb/flat/pdb2ltw.ent',
+            'mtime': time.time(), 'comment': None,
+        },
+    ]
+    mock_dir_crawler = Mock()
+    mock_dir_crawler.crawl.return_value = entries
+    mock_dir_crawler.__name__ = 'MockDirCrawler'
+    databank = DATABANKS[1]
+    databank['crawler'] = mock_dir_crawler
+
+    crawl(databank)
+
+    mock_storage.db.entries.bulk_write.assert_called_once()
+    mock_storage.db.entries.delete_many.assert_has_calls([
+        call({'databank_name': 'pdb', 'pdb_id': {'$nin': ['1crn', '2ltw']}})
+    ])
